Mueve Spinner a otro archivo para reutilizarlo en inicio
